Fix address size issue

+++ test/simplememorymanager/main.cc
@@ -69,7 +69,7 @@ void TestPiiSimpleMemoryManager::externalBuffer()
 {
   void* bfr = malloc(200); // Will not be freed if a test fails, but who cares?
   PiiSimpleMemoryManager manager(bfr, 200, 50);
-  QCOMPARE(manager.blockSize(), ulong(64 - sizeof(void*)));
+  QCOMPARE(manager.blockSize(), size_t(64 - sizeof(void*)));
   for (unsigned i=0; i<manager.blockCount(); ++i)
     QVERIFY(manager.allocate(50));
   QVERIFY(!manager.allocate(50));
